Feature 6/farmclub 기본 api

+++ src/main/java/com/modernfarmer/farmusspring/domain/farmclub/dto/res/GetMyFarmClubResponseDto.java
@@ -15,7 +15,7 @@ public record GetMyFarmClubResponseDto(
         Long wholeMemberCount,
         List<Step> steps,
         String advice,
-        int daysSinceStart
+        Long daysSinceStart
 ) {
 
     public static GetMyFarmClubResponseDto of(GetMyFarmClubVo farmClubInfo, List<Step> steps, String advice) {
